Allow lowercase propositions
Now propositions can be either case, while variables have to be
lowercase, and predicates have to be uppercase. In order for the
parser to error on nonsense such as

    ∃xp(x)

which otherwise would parse "(x)" as a parenthesised predicate, the
parser enters a "FOL" mode when parsing quantified formulas where it
does not recognize propositions.